moved plaid Link component to ConfigFrame module
